pyproject.toml
[project]
name = "veaiops"
version = "0.1.0"
description = "Volc Engine AIOps"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "aiocache>=0.12.3",
    "alibabacloud-cms20190101>=1.0.0",
    "alibabacloud-credentials>=1.0.0",
    "bcrypt>=4.3.0",
    "beanie>=2.0.0",
    "scipy>=1.15.3",
    "dbscan1d>=0.2.2",
    "fastapi>=0.116.1",
    "google-adk==1.15.1",
    "httpx>=0.28.1",
    "json-repair>=0.50.1",
    "loguru>=0.7.3",
    "lark-oapi>=1.4,<2.0",
    "markdownify>=1.2.0",
    "pillow>=11.3.0",
    "pillow-heif>=1.1.0",
    "pydantic-settings>=2.10.1",
    "python-jose>=3.5.0",
    "pyzabbix>=1.3.1",
    "starlette-context>=0.4.0",
    "tenacity>=8.5.0",
    "tos>=2.8.7",
    "urlextract>=1.9.0",
    "uvicorn>=0.35.0",
    "volcengine>=1.0.201",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-pymongo>=0.48b0",
    "tldextract>=5.3.0",
    "veadk-python[extensions]==0.2.18",
]

[project.optional-dependencies]
algorithm = [] # todo: add algorithm dependence

[tool.setuptools.packages.find]
include = ["veaiops*"]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.13.0",
    "tomli>=2.2.1",
    "types-cachetools>=6.2.0.20250827",
    "types-tqdm>=4.67.0.20250809",
    "bandit>=1.8.6",
]
docs = [
    "mkdocs-material>=9.6.19",
    "mkdocs-static-i18n>=1.3.0",
    "mkdocstrings[python]>=0.30.0",
]
test = [
    "mongomock-motor>=0.0.36",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.15.0",
    "pytest-cov>=4.0.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "frontend",
]
per-file-ignores = { "tests/**/*.py" = ["D"] }

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "D",   # pydocstyle
]
ignore = [
    "D100",  # Missing docstring in public module
    "D104",  # Missing docstring in public package
    "D105",  # Missing docstring in magic method
    "D107",  # Missing docstring in __init__
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
# Black-compatible import sorting
force-single-line = false
combine-as-imports = true
split-on-trailing-comma = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.codespell]
ignore-words-list = "medias"
